// src/app/components/formato-peritaje/interfaces/items-vehiculo.ts
export interface ItemsVehiculosUsados{
    id_veh_chk_usados: number;
    califica: string;
    fecini: Date;
    id_usuario: number;
    id_usu_inspector: number;
    id_cot_item_lote: number;
    prueba_ruta: number;
    fecfin: Date;
    califica2: number;
    id_veh_chk_usados_mas: number;
    llave: number;
    libro: number;
    cambio_correa: number;
    lugar_cambio_correa: string;
    herramienta: number;
    manual: number;
    reclamo_aseg: number;
    valor_reclamo_aseg: number;
    marca_radio: string;
    marca_llantas: string;
    marca_bateria: string;
    lugar_mante: string;
    cilindraje: string;
    color: string;
    referencia: string;
    clase: number; // Que es Clase?
    combustible: number;
    lugar_matricula: number;
    capacidad_sillas: number;
    km: number;
    fec_prox_mantenimiento: Date;
    cojineria: number;
    caja: number;
    traccion: number;
    rin: number;
    estribo: number;
    vidrios: number;
    techo_vidrio: number;
    retrovisores_electricos: number;
    no_airbags: number;
    no_rin: number;
    barra_techo: number;
    exploradora: number;
    placa: string;
    linea: string;
    marca: string;
    modelo: string;
}